// src/apis/models/FireBytes_Unified_Wave_BuildingFloors_Dtos_BuildingFloorEditDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_BuildingFloors_Dtos_BuildingFloorEditDto = {
    id?: string | null;
    floor?: string | null;
    description?: string | null;
    floorDisplayName?: string | null;
    latitude?: number | null;
    longitude?: number | null;
    pointXvalue?: number | null;
    pointYvalue?: number | null;
    pointZvalue?: number | null;
    border?: number | null;
    buildingId?: string | null;
    graphicPanelId?: string | null;
    graphicType?: string | null;
};